requirements.txt
defusedxml==0.7.1
fonttools==4.56.0
fpdf2==2.8.2
lxml==5.3.1
pillow==11.1.0
python-docx==1.1.2
striprtf==0.0.28
typing_extensions==4.12.2
